only the single processor version is currently working for empty polygons.

+++ src/openclimategis/profiling.py
@@ -3,7 +3,8 @@ import datetime
 import os
 from osgeo import ogr
 import warnings
-from util.ncconv.experimental.ocg_converter import ShpConverter
+from util.ncconv.experimental.ocg_converter import ShpConverter, KmlConverter,\
+    CsvConverter, LinkedCsvConverter
 import cProfile
 import pstats
 import sys
@@ -59,7 +60,7 @@ def f(polygons):
                                                    calendar='proleptic_gregorian',
                                                    time_units='days since 1950-01-01 00:00:0.0'),
                                      polygons=polygons,
-                                     time_range=[datetime.datetime(2011,11,1),datetime.datetime(2061,12,31)],
+                                     time_range=[datetime.datetime(2011,11,1),datetime.datetime(2012,12,31)],
                                      level_range=None,
                                      clip=True,
                                      union=True,
@@ -67,10 +68,11 @@ def f(polygons):
                                      max_proc=8,
                                      max_proc_per_poly=2)
         assert(sub.value.shape[2] > 0)
-        shp = ShpConverter(sub,'sresa1b_miroc3-2-medres_2_Prcp')
-        shp.convert(None)
+        csv = LinkedCsvConverter(sub,'foo')
+        out = csv.convert(None)
 
-def analyze():
+def analyze(drop=False):
+    
     engine = create_engine('sqlite:////home/bkoziol/tmp/profiling.sqlite')
     metadata = MetaData(bind=engine)
     Base = declarative_base(metadata=metadata)
@@ -120,44 +122,45 @@ def analyze():
         scenario = relationship(Scenario)
         function = relationship(Function)
         
-    metadata.drop_all(checkfirst=True)
-    metadata.create_all()
+    if drop: metadata.drop_all(checkfirst=True)
+    metadata.create_all(checkfirst=True)
     
     s = Session()
-    with open('/tmp/foo.txt','r') as out:
-        data = out.read()
-    profiles = re.split('finished ::.*',data)
-    profiles = profiles[0:-1]
-    for profile in profiles:
-        profile = profile.strip()
-        scenario_name = re.match('starting :: (.*)',profile).group(1)
-        scenario = Scenario.get_or_create(s,dict(name=scenario_name))
-        table = re.match('.*lineno\(function\)(.*)',profile,flags=re.DOTALL).group(1).strip()
-        lines = re.split('\n',table)
-        for line in lines:
-            line = line.strip()
-#            print line
-            elements = re.split('  {2,}',line)
-            if '{' in line and '}' in line:
-                filename = None
-            else:
-                filename_name = re.match('.* (.*):.*',elements[4]).group(1)
-                filename = FileName.get_or_create(s,dict(name=filename_name))
-            rm = re.match('.*\((.*)\)|.*{(.*)}',elements[4])
-            if rm.group(1) is None:
-                function_name = rm.group(2)
-            else:
-                function_name = rm.group(1)
-            function = Function.get_or_create(s,dict(name=function_name))
-            obj = Profile()
-            obj.ncalls = elements[0]
-            obj.tottime = elements[1]
-            obj.percall = elements[2]
-            obj.cumtime = elements[3]
-            obj.filename = filename
-            obj.scenario = scenario
-            obj.function = function
-            s.add(obj)
+    for kwds in data:
+        with open(kwds['name'],'r') as out:
+            data = out.read()
+        profiles = re.split('finished ::.*',data)
+        profiles = profiles[0:-1]
+        for profile in profiles:
+            profile = profile.strip()
+            scenario_name = re.match('starting :: (.*)',profile).group(1)
+            scenario = Scenario.get_or_create(s,dict(name=scenario_name))
+            table = re.match('.*lineno\(function\)(.*)',profile,flags=re.DOTALL).group(1).strip()
+            lines = re.split('\n',table)
+            for line in lines:
+                line = line.strip()
+    #            print line
+                elements = re.split('  {2,}',line)
+                if '{' in line and '}' in line:
+                    filename = None
+                else:
+                    filename_name = re.match('.* (.*):.*',elements[4]).group(1)
+                    filename = FileName.get_or_create(s,dict(name=filename_name))
+                rm = re.match('.*\((.*)\)|.*{(.*)}',elements[4])
+                if rm.group(1) is None:
+                    function_name = rm.group(2)
+                else:
+                    function_name = rm.group(1)
+                function = Function.get_or_create(s,dict(name=function_name))
+                obj = Profile()
+                obj.ncalls = elements[0]
+                obj.tottime = elements[1]
+                obj.percall = elements[2]
+                obj.cumtime = elements[3]
+                obj.filename = filename
+                obj.scenario = scenario
+                obj.function = function
+                s.add(obj)
     s.commit()
 
 if __name__ == '__main__':
